+++ examples/counter/src/app/app.component.spec.ts
@@ -1,14 +1,27 @@
 /* tslint:disable:no-unused-variable */
 
 import { TestBed, async } from '@angular/core/testing';
+import { Observable } from 'rxjs/Observable';
 import { AppComponent } from './app.component';
 import { CounterInfoComponent } from '../counter/counter-info.component';
 import { CounterComponent } from '../counter/counter.component';
 import { SearchComponent } from '../search/search.component';
+import { SelectByPathComponent } from '../select-by-path/select-by-path.component';
+import { SelectRootStateComponent } from '../select-root-state/select-root-state.component';
 import { NgRedux, DevToolsExtension } from 'ng2-redux';
 import { CounterActions } from '../counter/counter.actions';
 import { SearchActions } from '../search/search.actions';
 import { RandomNumberService } from '../common/random-number.service';
+
+const mockRedux = {
+  dispatch(action) {},
+  configureStore() {},
+  select() {
+    return Observable.from('test');
+  }
+};
+
+NgRedux.instance = mockRedux;
 
 describe('AppComponent', () => {
   beforeEach(() => {
@@ -18,9 +31,11 @@ describe('AppComponent', () => {
         CounterInfoComponent,
         CounterComponent,
         SearchComponent,
+        SelectByPathComponent,
+        SelectRootStateComponent,
       ],
       providers: [
-        NgRedux,
+        { provide: NgRedux, useValue: mockRedux },
         DevToolsExtension,
         CounterActions,
         SearchActions,
@@ -40,6 +55,6 @@ describe('AppComponent', () => {
     let fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     let compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector('h1').textContent).toContain('Counter Example');
+    expect(compiled.querySelector('h1').textContent).toContain('Counter Demo');
   }));
 });