+++ examples/counter/src/app/app.module.ts
@@ -8,6 +8,8 @@ import { AppComponent } from './app.component';
 import { CounterInfoComponent } from '../counter/counter-info.component';
 import { CounterComponent } from '../counter/counter.component';
 import { SearchComponent } from '../search/search.component';
+import { SelectByPathComponent } from '../select-by-path/select-by-path.component';
+import { SelectRootStateComponent } from '../select-root-state/select-root-state.component';
 import { RandomNumberService } from '../common/random-number.service';
 
 @NgModule({
@@ -21,6 +23,8 @@ import { RandomNumberService } from '../common/random-number.service';
     CounterInfoComponent,
     CounterComponent,
     SearchComponent,
+    SelectByPathComponent,
+    SelectRootStateComponent,
   ],
   bootstrap: [ AppComponent ],
   providers: [