+++ examples/counter/src/select-by-path/select-by-path.component.ts
@@ -1,0 +1,16 @@
+import { Component } from '@angular/core';
+import { select } from 'ng2-redux';
+import { Observable } from 'rxjs/Observable';
+
+/**
+ * A component demonstrating selecting deeply nested store properties
+ * with the path syntax.
+ */
+@Component({
+  selector: 'select-by-path',
+  templateUrl: './select-by-path.component.html',
+})
+export class SelectByPathComponent {
+  @select([ 'pathDemo', 'foo' ]) foo$: Observable<Object>;
+  @select([ 'pathDemo', 'foo', 'bar', 0 ]) bar$: Observable<number>;
+}